Support the X2APIC MADT ACPI table entry (type 9) (#866)
* Haven't yet found any systems that use this,
  but if we do, it is now supported.

* Also, all processor ID and APIC ID fields from MADT
  table entries are now converted to `u32`.

+++ kernel/ap_start/src/lib.rs
@@ -35,7 +35,7 @@ pub fn insert_ap_stack(cpu_id: u32 , stack: Stack) {
 /// Entry to rust for an AP.
 /// The arguments must match the invocation order in "ap_boot.asm"
 pub fn kstart_ap(
-    processor_id: u8,
+    processor_id: u32,
     cpu_id: CpuId,
     _stack_start: VirtualAddress,
     _stack_end: VirtualAddress,
